refactor(loans): rename component file to follow naming conventions
